Redone Pod and Art Assets
I've redone the mining pod from scratch in a level called
"levelCrater_Zach."  Improvements include:
- A nice clean new event graph with more streamlined functions
- The pod now pitches and rolls as the player moves
- Upgraded to a skeletal mesh/the collector units on the sides move up
and down with the mouse
- The firing explosion effect now moves up and down with the collector
units
- The camera lags slightly behind the pod, giving it a more dynamic feel
- The LMB/LT is now the only fire button.  InvLeft fires first.  RMB/RT
now controls zoom.
- I also replaced the projectile mesh with the ore vein mesh.

Additionally, the new MiningPod.fbx file is in props, and I added the
pipe FBX files into a subfolder of Props called "Pipes."  Two files
called mining_pod_v2.ma and pipeset.ma are in the Art Dropbox folder.
Both are fully UVd and can be used as a reference to texture the pod and
pipes.